requirements.txt
requests==2.31.0
Flask==3.0.0
PyYAML==6.0.1
beautifulsoup4==4.12.2
